:recycle: separete util.go to server and client
